requirements.txt
accelerate==0.34.2
ago==0.0.95
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argparse==1.4.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.1.0
boto3==1.34.40
botocore==1.34.40
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.2
cssselect==1.2.0
cycler==0.12.1
datasets==3.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.8.0
elastic-transport==8.12.0
elasticsearch==8.12.0
executing==2.0.1
fastapi==0.114.2
fastjsonschema==2.19.1
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gguf==0.9.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hf_transfer==0.1.6
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
incremental==22.10.0
interegular==0.3.3
ipykernel==6.29.4
# ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langdetect==1.0.9
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistral_common==1.4.0
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
news-please==1.5.44
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.45.0
outlines==0.0.46
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.8.2
pexpect==4.9.0
pillow==10.4.0
plac==1.4.2
platformdirs==4.2.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==24.6.1
pycparser==2.21
pydantic==2.9.1
pydantic_core==2.23.3
PyDispatcher==2.0.7
Pygments==2.17.2
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
ray==2.35.0
readability-lxml==0.8.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
s3transfer==0.10.0
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.12.0
Scrapy==2.11.0
Send2Trash==1.8.3
sentence-transformers==3.0.0
sentencepiece==0.2.0
sentry-sdk==1.45.0
service-identity==24.1.0
setproctitle==1.3.3
setuptools==68.2.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.38.5
sympy==1.12
terminado==0.18.1
threadpoolctl==3.3.0
tiktoken==0.7.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.44.1
triton==3.0.0
Twisted==22.10.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.30.6
uvloop==0.20.0
vllm==0.6.1.post2
vllm-flash-attn==2.6.1
w3lib==2.1.2
wandb==0.16.6
warcio==1.7.4
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==13.0.1
wheel==0.41.2
widgetsnbextension==4.0.10
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
zope.interface==6.1